backend/requirements.txt
# Google Sheets API dependencies
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AWS dependencies
boto3==1.34.0
botocore==1.34.0

# Standard dependencies
requests==2.31.0